Bug#29288023: Join condition is substituted with REF access
In regular MySQL execution, materialized derived tables may be augmented
with generated indexes that match predicates from the actual query and
which are deemed useful during query plan generation.

Currently, secondary engines do not support any kind of indexes,
so generating metadata for indexes on such tables is not useful.
In addition, REF access patterns generated for such indexes cause the
corresponding equality predicate to be removed from the query plan,
thus generating a proper access plan might be more difficult for the
secondary engine.

This patch ensures that no indexes are generated for materialized derived
tables when a secondary storage engine is targeted.

Also ensure that we have proper error handling, separating an error return
from a decision not to create an index.

Reviewed by: Dag Wanwik <dag.wanvik@oracle.com>