Bug#29411899: CAN'T UPGRADE FROM 5.7 TO 8.0 IF ANY DATABASE HAS A HYPHEN IN THEIR NAME
An FTF table was created in a database in which the db_name contained a hyphen.
This was done in 5.6, then upgraded to 5.7, then to 8.0.  The problem was that
there was an entry in the dd:tablespaces but not in dd:tablespace_files.
This caused the file to not be found because the filepath generated used the
hyphen instead of "@002" which is what the server changed the hyphen to
in the identifier.

The fix is to modify dict_check_sys_tables() so that it checks if the filepath
is actually found in the DD and if not, makes it correctly.

Approved by Rahul Agarkar in RB#21650