Bug#28940167 - WAIT_TIMEOUT ERROR NOT CLEAR AND NOT SENT TO CLIENT BEFORE CLOSING CONNECTION
Problem:
If the clients are not using the connection within wait_timeout, then they will
be closed by the server and the server can write a non-clear error message to
the error log:
[Note] Aborted connection <connection_id> to db: '<db>' user: '<user>' host:
'<host>' (Got timeout reading communication packets)

Which also is the same as if there is an error due to net_read_timeout.

And the client is not informed of the reason for the closed connection, it
typically sees:
2013: 'Lost connection to MySQL server during query'
or
2006: 'MySQL server has gone away'

These are very generic errors and does not give a clear hint that the
connection was closed by the server due to wait_timeout.

Solution:
The 'wait_timeout' is now mentioned in the server error log as a
separate error(note) if it triggered the aborted connection. New error message
introduced ER_NET_WAIT_ERROR. There was also plans to introduce this
error message to the client before it disconnects. This will be done
by WL12999.

Thanks to Mattias Jonsson mattias.jonsson@booking.com for suggesting this.

Change-Id: Ib3f98d1b76383d5b9d4921f4f3eba83baad9c72c