BUG#29111514 FLUSHING BINARY LOGS WITH GTID LEADS TO AN OOM ERROR
Description
-----------
The `binlog_gtid_table_readonly.test` script is failing on Windows platform
when ran repeatedly.

Analysis
--------
The count for the amount of lines, on the error log, that match the expected
errors is more than the accounted for.

Fix
---
Replace the `assert_count` with the `assert_match` parameter for the
`assert_grep.inc` script invocations.

ReviewBoard: 21628
Reviewed-by: Neha Kumar <neha.n.kumari@oracle.com>