Bug#29436791 MYSQL DOES NOT COMPILE WITH PROTOBUF 3.7.0
Building with libprotoc 3.7.0 gives lots of warnings:
warning: 'PROTOBUF_MIN_PROTOC_VERSION' is not defined, evaluates to 0 [-Wundef]

The fix is to do
ADD_DEFINITIONS(-DPROTOBUF_MIN_PROTOC_VERSION=2006000)

Building with libprotoc 3.7.0 breaks the build for some unit tests.
Classes generated by protoc are tagged as 'final' so we cannot derive
from them.

The fix is to change class Message_from_str from 'is-a' Message_type
to 'has-a' Message_type.

Change-Id: I1a160e8e3b9de60e0aeecfece94cbd277a118211