Bug#29328158: UNDO FILE PATH IN DD NOT UPDATED IF THE UNDO FILES MOVED TO A DIFFERENT FOLDER
When an explicit undo tablespace is moved to a different valid directory
that is scanned at startup, the new filepath is not updated in the DD as
reflected by information_schema.files.

Adjusted Validate_files::check() to do for undo tablespaces what it does
for ibd files to update the DD if they are moved.

Approved by Rahul Agarkar in RB#21606