Bug#29389389 SPORADIC FAILURE TO BUILD MYSQL-5.6
Build for cluster mysql-5.6 failing with error :
"ISO C++ forbids comparison between pointer and integer [-fpermissive]"
as the code was not compatible with higher version of gcc [gcc8] than
the recommended version of gcc for 5.6.

Hence, the line of code is changed to match that of versions higher
than 5.6 to adapt to the newer gcc versions and has been tested for
the recommended or lower gcc versions as well.

Reviewed by: Georgi Kodinov <georgi.kodinov@oracle.com>