Bug#29454074 COMPONENT TESTS ON WINDOWS SLOW DUE TO NON-BLOCKING SOCKET IN TCPPORTPOOL
The TcpPortPool is the mtr-compatible way of router component tests using
system ports. We pick the port based on the availability  based on
existence of the lock files.

Additionally there is a mechanism to check if the port is really available by
trying to connect to it. On Windows this probe takes about 1s for each port
we check this way. This creates quite significant time overhead for tests
running on windows.

This patch sets the socket to non-blocking before testing port with the connect()

RB: 21686
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>