WL#12303 UPDATE CLUSTER COMPILE SCRIPTS TO USE CMAKE3
Alter compile-cluster script to allow usage of 'cmake3', if it exists
or 'cmake' or stop script execution with appropriate message if both
do not exist.

All machines used by autotest to execute tests use cmake3 for higher
versions of cmake3 and there is a possibility of the cmake command
being deprecated. Hence, the patch gives room for even lower versions
of mysql-cluster to use cmake3 if cmake is deprecated.

Reviewed by: Tiago Alves <tiago.alves@oracle.com>