Bug#29153009 - ENSURE SEED OPENS PORT BEFORE CREATING REMAINING NODES IN GCS SMOKE TESTS
Description:
-----------
The GCS multinode smoke tests, in particular those that create lots of
processes such as SendMessageTest_NoOp_Group or BootTest_NoOp_Group, may
produce false negatives due to their implementation.
Essentially the implementation assumes that when it creates the non-seed
nodes, the seed has already initialized and is listening on its port.
The tests usually do this by waiting for some time after creating the seed
before creating the remaining nodes (e.g. SendMessageTest_NoOp_Group),
sometimes waiting no time at all (e.g. BootTest_NoOp_Group).
Basically, the implementation "hopes" that the seed initializes before the
remaining nodes exhaust their connection attempts to the seed.

If the assumption above does not hold, we have a false negative when a test
fail. It is extremely annoying to spend time analysing what turns out to be a false
negative.

How To Repeat:
-------------
Execute the GCS multinode smoke tests.

Fix
------------

Created a test util that identifies if a port is busy. With that, one did:
- Create an utility to hand out addresses for tests, to avoid port
clashes.
- Augment the wait_for_start method to wait for a port to become really
busy, in order to proceed with the test.

RB: 21448