Bug #29440408 ASSERTION FAILURE: BUF0BUF.CC:4040:M_MODE == PAGE_FETCH::POSSIBLY_FREED || !BLOC
Problem:

When a new record is inserted, it is possible that this operation is carried
out by doing an update of a delete-marked record.  If the server crashes after
undelete marking of the record, but before inserting and updating the blob
reference, the undeleted record will actually contain stale blob reference
value.  Since the blob insert does log_free_check() periodically, it is
possible for the purge thread to actually free LOB via the undo record.
During recovery, because of the stale value of LOB reference, the same LOB
will be attempted to be freed again (but now via the newly inserted record).
This results in access of the freed LOB page.

Solution:

Don't allow the stale blob reference in the undelete marked record.  If
the LOB operation code is OPCODE_INSERT_UPDATE, then clear out the
page number in the LOB reference.

rb#21699 approved by Rahul Agarkar.