Bug#29471723 BROKEN BUILD FEDORA 28, UNDEFINED REFERENCE TO SYMBOL '__BSS_START'
The patch for
Bug#29389789 HANDLE SIGNALS, AND ENABLE STACK TRACE IN UNIT TESTS
added dependencies on stack tracing routines from gunit_test_main.cc.

On Fedora 28, we get undefined reference to symbol '__bss_start' for
several unit tests and plugins when linking with the
standard linker (not using -fuse-ld=gold)

The workaround is to add "--copy-dt-needed-entries" to linker flags.

Change-Id: Ib07f03874e795039719eb82a02ea695fb8e8c82c