Bug#29257254: SIG11 IN REINIT_STMT_BEFORE_USE() FOR PREPARED STATEMENT
Single-table UPDATE and DELETE statements that used indexed
expressions, could cause the server to exit when executed as prepared
statements.

When an indexed expression is replaced with the corresponding indexed
generated column internally, a hidden column may be added to the
statement. Usually, the hidden column is removed before the next
execution of the statement. However, single-table UPDATE and DELETE
statements have a fast path that skips some preparation and
optimization steps, and the removing of the hidden column is part of
the skipped code. The result is that each execution of the prepared
statement adds a new hidden column to the internal data structures,
until the list of columns becomes too big.

Fixed by resetting the column list also at the beginning of
single-table UPDATE and DELETE statements.

Change-Id: I24f81d708786c7a194f2d00cfff135a5f68f6836