BUG#29264874: ASAN ERROR WHEN ACCESSING INFORMATION SCHEMA               TABLES FROM AGGREGATE UDF
ANALYSIS:
=========
With ASAN, a SELECT on I_S tables using aggregate UDF
functions reports a memory corruption error.

When Information Schema is queried for table statistics,
we access 'dd::tablespaces::se_private_data' and
'dd::Table::se_private_data'. In the current scenario,
these fields are not populated. Memory is allocated for
them, but they are not initialized. Hence when we try to
access them in 'Field_blob::val_str()', it results in an
undefined behavior.

FIX:
====
Field::pack_length() number of bytes are now initialized
when we open a table based on table definition in
'open_table_from_share()'.

ReviewBoard: 21697