Bug#29428288 CRASH IN PREPARE_BLOB_FIELD()
When a user executes ALTER TABLE ... CONVERT TO CHARACTER SET,
the server goes through all existing columns in the given table
to see if any of the data types needs to be changed in order to
fit the existing data (going from a single-byte to multi-byte
character set for instance). This check includes looking at the
character set for the old column definition. If the user also did
an "ADD COLUMN"-operation in the same ALTER TABLE statement, the
server would end up doing the same check for this new column. The
result was that the server ended up dereferencing a null-pointer
when trying to look at the old column definition for the new
column. This patch avoids doing this checking for the new column
since there is no old column definition to look at.

Change-Id: Ibd58fc30caa21e6bc61aedab835b8291f5201daa