Bug #25945005: SERVER DOES NOT CHECK CERTIFICATE EXPIRY RB#21716
There were no checks done on the validity of the server certificate.
E.g. the validity of it wasn't checked. But that's just an edge case
of checking the integrity of it.
Luckily openssl provides a way to validate this certificate better.
This fix adds a call to that validation for the server listening
socket and produces a warning in the server log if it doesn't succeed.
It could be promoted to an error too, but there are legitimate cases
when the certificate doesn't "verify" against the local CA etc but is
still "valid" as it verifies against the CAs supplied by the client.
Thus turning this to an error right away would be an incompatible
change.

Unfortunately wolfSSL doesn't implement all APIs needed for that.