Bug#23512140 - CODE IN GCS_XCOM_PROXY::XCOM_WAIT_* IS DUPLICATED
Description
-----------
The functions in GcsXcomProxy that wait for XCom events implement duplicated
code. This code is responsible for setting up and executing a timed_wait, as
well as for the corresponding error checking.

Solution
--------
Isolate the duplicated code in a generic 'wait_for_condition' function to be
called by the event specific functions.

RB: 21391
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed-by: Filipe Campos <filipe.campos@oracle.com>
Reviewed-by: Tiago Vale <tiago.vale@oracle.com>