Bug #29486219: RECENT REGRESSION: CRASH IN FILESORT::MAKE_SORTORDER()
When setting up filesort for DISTINCT, turn off the transformation of
BIT fields to LONG; we don't need it (it's only needed for certain kinds
of temporary tables), and it somehow causes items to be pointing into the
wrong slice.

Longer term, we should remove this transformation when we remove the
MEMORY storage engine, but it's solidly entrenched all around the server,
so it won't necessarily be easy.

Change-Id: I2b68cfcbd9d2a1cbf3143e186e1390deb08d04d2