Bug#29258979: ASSERTION FAILED: CS_NUMBER < ARRAY_ELEMENTS(ALL_CHARSETS)
get_collation_number_internal() uses collation name to get its collation
number. We use std::unordered_map::operator [] to do it. But if user uses
an invalid collation name, std::unordered_map::operator [] inserts a new
key-value pair to the map. The insertion is not thread-safe, which causes
this assertion failure.

Fix:
Check the existence of a collation when getting its collation number.

Change-Id: Id1ca80f74b4283fac23752b5fe36d50b03539e18