Bug #29507630: MEMBER FUNCTION SHOULD NOT BE NAMED MAIN()
The X plugin has a member function called main(). This is a big no-no; it
breaks that "break main" in gdb will put a breakpoint on that function in
addition to the regular system main(). This, in turn, means that mtr --gdb
_always_ stops in the X plugin, which is not what most developers want.

Thus, we rename the function to plugin_main(). Rename exit() while we're at it.

Change-Id: Ic39cd4950c6cf3aaf62b22d5673b895198632551