Bug#27675107: SUPER_READ_ONLY WITH MASTER_INFO_REPOSITORY=TABLE BREAKS               REPLICATION ON RESTART
Problem
-------
1. The table mysql.slave_master_info which is responsible for storing
the master specific information is not updated on shutdown when the
server is in super-read-only mode.

2. Any errors generated at the time of commit to the repository tables
were not propagated upwards causing no entries of errors in the error
log.

Analysis
--------
- Normally, the slave repository tables are updated by replication
  threads.
- Replication threads are excluded from read-only checks.
- On server shutdown, a non-replication thread tries to update the
  slave position tables. It uses a special THD object for the session,
  which is not excluded from the read-only check.
- Hence the positions are not persisted during server shutdown.

Fix
---
1. The shutdown thread should update the 'slave_master_info' table
even when super_read_only is set.  The 'skip_readonly_check' flag is
now enabled for the thread which tries to update the
'slave_master_info' the system table during shutdown.

2. Made the Rpl_info_table_access::close_table method return bool for
success/failure, and handled the error propagation in case of shutdown
of server when super_read_only mode is set on slave.

RB: 21729