Bug #29514504: WRONG RESULT WITH CORRELATED LATERAL JOIN
Fix an issue with derived tables with aggregation, where the aggregation
gets zero input rows; it would not properly copy the results of the
aggregate functions into the temporary table.

This would normally be fine, as the default value would be a good one,
but not if the derived table was evaluated multiple times (e.g. LATERAL).

Change-Id: I447ee454f89d8fcc8894c54c1fc679c27e28dddb