Bug #29420141: SET_PERSIST_OPTIONS DOES NOT VALIDATE USER NAME LENGTH/HOST NAME LENGTH RB#21762
The size of the incoming data from the generated file was not
checked when the host and user buffers were filled in.
Added size checks to copying the host and user names from
mysqld-auto.cnf.
Added a debug assert if the config name fits the provided buffer.
Added an debug assert to check the size of the destination host buffer
when reading from thd.