Bug#28702320 - INCORRECT GTID GENERATION WHEN IN PRESENCE OF MEMORY FAILURES
Description
-----------
This issue is caused by an incomplete handling of OOM errors in XCom.
When XCom sends a message, it copies the payload from the local data;
however, if the copy fails to allocate memory for the message, it will
simply print an error message, but will still send the message - a
corrupt one with null payload. This null message eventually gets to GCS
on the remote node, which detects that the message is empty and discards
it. The XCom instance at the remote node, however, is satisfied with the
response it gets, so it will not retry to get the correct version of the
corrupt message, effectively resulting in a skipped message.

Solution
--------
Terminate XCom when it gets an OOM error.

RB: 21644
Reviewed by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed by: Tiago Vale <tiago.vale@oracle.com>