Bug#29526078 FIELDS REFERRED FROM PUSHED CONDITIONS NOT REQUIRED TO BE IN THE READ_SET
Several Field::get'ters contain an assert that check for the Field
being part of the table read_set. This is only a valid assumption when
the Field get'ers are used in the servers 'execute query phase'.

Several places the same Field get ters are used in other context, like
in the handler code for processing row contents before it reach the
servers 'arena'. In these cases it is a more or less established pattern
to use 'dbug_tmp_use_all_columns()' - 'dbug_tmp_restore_column_map()' to
temporary allow 'all_columns' from a table to be referred, even if not
in the read_map originally.

This patch add the same pattern to get_field_val_str() which is used
by the ha_ndbcluster handler when generating pushed condition code.

Reviewed by: Magnus Blaudd <magnus.blaudd@oracle.com>