Bug#29531670: MYSQL_UPGRADE SHOULD RETURN 0 UNCONDITIONALLY WHEN BEING DEPRECATED
Since the server now executes the statements previously executed by
mysql_upgrade, the client is now deprecated, and returns the value 2, meaning
that the server is already upgraded. The intention of returning a value != 0
was to make users aware of the deprecation. However, there are indications
that this results in undesirable breakage of toolchains, and hence, we change
the client to return 0 unconditionally.

Change-Id: I91c83fb220bd42bebfda67cd152eb81ad54a6685