Bug#29501263 CRASH IN SPJ BLOCK, ILLEGAL ACCESS TO ARRAYPOOL<T>::GETPTR
When the SPJ block called lookup_resume() to schedule an
operation previously enqueued, it used a 'correlation id' which
could have been produced from its immediate ancestor in the execution
order, and not its parent in the queryTree as assumed.

This patch checks whether the 'execution ancestor' is different from
the queryTree parent. If not the same, the correlation id from the
queryTreee parent is looked up and that correlation id enqueued to
be executed later.

Reviewed by: Maitrayi Sabaratnam  <maitrayi.sabaratnam@oracle.com>