Bug #29343655 NDB : NDB_RESTORE INCORRECTLY PRINTS STOPGCP-1 WHEN RESTORING BACKUP
From: Frazer Clement <frazer.clement@oracle.com>

Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>

Since time immemorial, data nodes have written StopGCP-1 into their
Backup control files.

ndb_restore corrects this by adding 1 when using StopGCP to control
log replay, and when restoring an 'epoch' entry into the
mysql.ndb_apply_status table.

However it does not correct it when printing the StopGCP of the
backup to stdout during execution.

This is fixed in this patch by correcting the problem when loading the
Backup metadata, so that it is set to the correct value for all 3
use cases in ndb_restore.

The erroneous use of -1 in the data node is not modified as that would
represent an incompatible backup format change.