WL#11999 Add support for JSON Schema
JSON_SCHEMA_VALID depends on RapidJSON patches that are not part of
the most recent released version, which is 1.1.0 at the time of
writing. Enhance the version check for WITH_RAPIDJSON=system in CMake
to also check that the system version of RapidJSON contains the
necessary fixes.

Also stop adding the WITH_RAPIDJSON=system build flag automatically
when WITH_SYSTEM_LIBS is specified, since it's not likely that the
system version of RapidJSON contains these fixes until they are part
of a released version.

Change-Id: I8747685472d957baf686f52fd9d633fb6d0b0dcf