WL#10162 Test participant fail during schema operation
Add test to verify what happens if a MySQL Server leaves(fail,
disconnect or unsubscribe) during a schema operation.

With the new schema operation result, the coordinator will add result
describing that a participant left and the result should be visible as
warnings in the client when the schema operation completes.

Change-Id: I604f3231732af9282bee744400a504ea6cd912c4