Bug#28834208 row_log_table_apply_update*inplace_alter_table_impl*assertion failure:row0log.cc
Problem:
The table has an index on the virtual generated column.
If the optimization table is in progress, and there is an update
on the table which modifies some base column of a virtual column
but not resulting in the virtual column value change.
Then it see index entry not found while applying these updates
on the newly created table as part of the optimize table.

Analysis:
As part of log apply on newly table,
InnoDB is not populating the virtual column data
before checking if it required to update the index or not.

Fix:
Populate the virtual data and then check if it required
to update the index.

RB#21730
Approved by Bin Su<bin.x.su@oracle.com>