Bug#29550442: COMP_ERR READS UNINITIALIZED DATA
comp_err calls my_strtoll10() with an uninitialized end argument. This
results in reading of uninitialized data when my_strtoll10() checks if
it has reached the end of the string.

Fixed by passing nullptr instead an uninitialized end pointer. The end
argument is only needed if my_strtoll10() should stop before the
terminating null character, or if the caller wants it to return a
pointer to the end of the parsed number. It is not needed in this
case.

Change-Id: I026b52e18f2bdf78cd7838d89fdbaa13a0a0acc5